Feat : 일과 UI 정보 표시
- 일과 진행 창(진행 텍스트, 애니메이션, 진행바)
- 일과 결과 창(변경된 능력치 표시, 결과 텍스트)
- 일과창 클릭 시 헬스, 가만이 있기 일과 자동 선택

+++ Assets/03.Scripts/Managers/DataManager/PlayerData.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 [Serializable]
 public class PlayerData
 {
-    public float PlayerGold = 0;
+    public int PlayerGold = 0;
     public SerializedDictionary<Define.PlayerStatType, int> Stats = new SerializedDictionary<Define.PlayerStatType, int>();
     public int ChoiceNumber = -1;
